requirements.txt
Django>=5.0.6,<5.1
djangorestframework>=3.15.1,<3.16
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.27.2,<0.28
pgvector>=0.2.5,<0.3
langchain>=0.2.3,<0.3
langchain-community>=0.2.4,<0.3
langchain-huggingface>=0.0.3,<0.1
accelerate>=0.31.0,<0.32
langchain_postgres>=0.0.8,<0.1
langchain-nomic>=0.1.2,<0.2
tiktoken>=0.7.0,<0.8
langchainhub>=0.1.20,<0.2
nomic[local]>=3.0.33,<3.1
chromadb>=0.5.3,<0.6
langgraph>=0.0.69,<0.1
tavily-python>=0.3.3,<0.4
langchain-text-splitters>=0.2.1,<0.3
beautifulsoup4>=4.12.3,<4.13
lark>=1.1.9,<1.2